Small changes in the tests

+++ src/test/java/edu/kit/informatik/pcc/service/videoprocessing/chain/decryption/DecryptTest.java
@@ -45,9 +45,9 @@ public class DecryptTest {
 
     @After
     public void cleanUp() {
-        /*if (decVid.exists())
+        if (decVid.exists())
             decVid.delete();
         if (decMeta.exists())
-            decMeta.delete();*/
+            decMeta.delete();
     }
 }
